convert dataclass to pydantic
